refactor: rename unauth to globals
