+++ src/app/operations/globals.rs
