+++ src/app/operations/mod.rs
@@ -1,2 +1,2 @@
 pub mod auth;
-pub mod unauth;
+pub mod globals;
